client.rs structs moved to new files for clarification
